storage_manager

Module catalog

Structs

Catalog

Column

Database

Table

Constants

CATALOG_DIR

CATALOG_FILE

DATABASE_DIR

DATA_DIR

TABLE_DIR_TEMPLATE

TABLE_FILE_TEMPLATE

Functions

create_database

create_table

Creates a new table under the specified database and updates the catalog on disk. If the table already exists, it will not be overwritten.

init_catalog

load_catalog

save_catalog

show_databases

Displays all databases available in the catalog. If no databases exist, it prints an appropriate message.

show_tables