Trait storage_manager::StorageDevice

pub trait StorageDevice: Downcast {
    fn read_sectors(
        &mut self, 
        buffer: &mut [u8], 
        offset_in_sectors: usize
    ) -> Result<usize, &'static str>;

    fn write_sectors(
        &mut self, 
        buffer: &[u8], 
        offset_in_sectors: usize
    ) -> Result<usize, &'static str>;

    fn sector_size_in_bytes(&self) -> usize;

    fn size_in_sectors(&self) -> usize;

    fn size_in_bytes(&self) -> usize { ... }

    fn block_bounds(
        &self, 
        offset: usize, 
        length: usize
    ) -> Result<BlockBounds, &'static str> { ... }
}

A trait that represents a storage device, such as hard disks, removable drives, SSDs, etc.

It offers functions to read and write at sector granularity, as well as basic functions to query device info.

Required methods

fn read_sectors(
&mut self,
buffer: &mut [u8],
offset_in_sectors: usize
) -> Result<usize, &'static str>

Reads content from the storage device into the given buffer.

Arguments

buffer: the destination buffer. The length of the buffer governs how many sectors are read, and must be an even multiple of the storage device's sector size.

offset_in_sectors: an absolute offset from the beginning of the storage device, given in number of sectors. This is sometimes referred to as the starting logical block address (LBA).

Returns the number of sectors (not bytes) read into the buffer.

fn write_sectors(
&mut self,
buffer: &[u8],
offset_in_sectors: usize
) -> Result<usize, &'static str>

Writes content from the given buffer to the storage device starting at the given offset_in_sectors.

Arguments

buffer: the source buffer. The length of the buffer governs how many sectors are written, and must be an even multiple of the storage device's sector size.

offset_in_sectors: an absolute offset from the beginning of the storage device, given in number of sectors. This is sometimes referred to as the starting logical block address (LBA).

Returns the number of sectors (not bytes) written to the storage device.

fn sector_size_in_bytes(&self) -> usize

Returns the size of a single sector in bytes, as defined by this drive.

fn size_in_sectors(&self) -> usize

Returns the number of sectors in this drive.

Loading content...
Provided methods

fn size_in_bytes(&self) -> usize

Returns the size of this drive in bytes, rounded up to the nearest sector size. This is nothing more than sector_size_in_bytes()*size_in_sectors().

fn block_bounds(
&self,
offset: usize,
length: usize
) -> Result<BlockBounds, &'static str>

Calculates block-wise bounds based on a byte-wise offset and buffer length.

Arguments

offset: the absolute byte offset from the beginning of the block storage device at which the read/write starts.

length: the number of bytes to be read/written.

Return

Returns a BlockBounds object.

If offset + length extends past the bounds, the Range will be truncated to the last block of the storage device (all the way to the end), and the last_block_offset will be 0.

Returns an error if the offset extends past the bounds of this storage device.

Loading content...
Methods

impl dyn StorageDevice + 'static

pub fn is<__T>(&self) -> bool where
__T: StorageDevice,

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T>(
self: Box<dyn StorageDevice + 'static>
) -> Result<Box<__T>, Box<dyn StorageDevice + 'static>> where
__T: StorageDevice,

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_ref<__T>(&self) -> Option<&__T> where
__T: StorageDevice,

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T>(&mut self) -> Option<&mut __T> where
__T: StorageDevice,

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.