Crate storage_manager

Manages and handles initialization of all storage devices and storage controllers in the system.

Traits

StorageController

A trait that represents a storage controller, such as an AHCI controller or IDE controller.

StorageDevice

A trait that represents a storage device, such as hard disks, removable drives, SSDs, etc.

Functions

init_device

Attempts to handle the initialization of the given PciDevice, if it is a recognized storage device.

storage_controllers

Returns an iterator over all initialized storage controllers on this system.

storage_devices

Returns an iterator over all storage devices attached to the storage controllers on this system.

Type Definitions

StorageControllerRef

A trait object wrapped in an Arc and Mutex that allows arbitrary storage controllers to be shared in a thread-safe manner.

StorageDeviceRef

A trait object wrapped in an Arc and Mutex that allows arbitrary storage devices to be shared in a thread-safe manner.