Trait storage_manager::StorageDevice

pub trait StorageDevice: BlockIo + BlockReader + BlockWriter + KnownLength + Downcast {
    fn size_in_blocks(&self) -> usize;
}

A trait that represents a storage device, such as hard disks, removable drives, SSDs, etc.

A StorageDevice must implement the following traits:

BlockIo: to specify its block size.

BlockReader and BlockWriter: to enable reading and writing data from/to the device at block granularity.

KnownLength: to specify the size in bytes (length) of the entire device.

This trait includes additional functions to query device info, e.g., the device’s total size in number of blocks.

Required Methods

fn size_in_blocks(&self) -> usize

Returns the total size of this device, given in number of blocks (sectors).

Implementations

impl dyn StorageDevice + 'static

pub fn is<__T>(&self) -> boolwhere
__T: StorageDevice,

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T>(
self: Box<dyn StorageDevice + 'static, Global>
) -> Result<Box<__T, Global>, Box<dyn StorageDevice + 'static, Global>>where
__T: StorageDevice,

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T>(
self: Rc<dyn StorageDevice + 'static>
) -> Result<Rc<__T>, Rc<dyn StorageDevice + 'static>>where
__T: StorageDevice,

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T>(&self) -> Option<&__T>where
__T: StorageDevice,

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T>(&mut self) -> Option<&mut __T>where
__T: StorageDevice,

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.